fixed the error with reading .bin file
On the 4-th round of reading the file in the meaningfull part of the packet exists two bytes with the value of 0x0C, which means '\f' symbol in C++., which is skipped by >> operator by default. So the data part, that should be 69 bytes, become 67 and two bytes was peaked from the 'size' part of the next package.
So I add 'noskipws' manipulator to avoid it.